Type changed from Long to Integer.

+++ src/main/java/com/erp/demo/repository/EmployeeInfoRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface EmployeeInfoRepository extends CrudRepository<EmployeeInfo, Long> {
+public interface EmployeeInfoRepository extends CrudRepository<EmployeeInfo, Integer> {
 
     List<EmployeeInfo> getAllByEmployeeIdAndDepartmentAndJobTitle(int employeeId, String department, String jobTitle);
     List<Integer> getEmployeeIdByActiveFlag(char activeFlag);
