+++ src/main/java/com/erp/demo/repository/TimesheetRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface TimesheetRepository extends CrudRepository<Timesheet, Long> {
+public interface TimesheetRepository extends CrudRepository<Timesheet, Integer> {
 
     List<Timesheet> getAllByEmployeeIdAndMonthAndFinancialYear(int employeeId, String month, int financialYear);
